save user metadata works on mobile
